[HW][IST] Drop debug printing from InnerSymbolTable.
A line is printed for every encountered symbol + its target,
often from IST's constructed in parallel producing a lot
of essentially unreadable debug information that rarely is useful.

Given IST's use in verification, this also means `-debug` quickly
becomes very noisy (between passes or when printing IR).

Drop for now, can add a dump() and/or PrintISTPass in the future.