[FIRRTL][Dedup] Change port renaming strategy
When we determine if two modules can be merged we ignore the names of
fields in bundle types.  When we merge two modules with different bundle
field names, this results in the type of instance results changing.  We
used to run a procedure to fix up subfield and connect operations on
instance results, but this missed some other operations which work with
bundle types. we now convert from the new type back to the old type
using an intermediate wire.