[Arc] RemoveUnusedArcArguments pass improvements (#4919)
Most importantly, this adds support for the new arc.call operation. Otherwise, it will not consider it and thus lead to errors because the argument list does not match anymore after the pass.

Since removing arguments of one arc can now lead to more unused arguments in another arc due to the removal of operands in the arc.call, it is changed to use the pattern rewriter to check for these cases again instead of manually keeping track of it.
I also added some pass statistics to see how many arguments are actually removed in how many arcs.