[Arc] Separate memory representation before and after state lowering (#4949)
Currently, the memory read and write operations represent the memory accesses both in the dataflow representation and the controlflow representation after LowerState. The semantics are different depending on whether LowerState was already performed or not and it does not make sense to have a clock operand after state lowering. Due to these differences it probably makes sense to have separate operations.
This commit also makes the clock operands of the dataflow variant optional such that it can be removed when inside a clock domain (analogous to the arc.state operation).