[Seq] Add enums for RUW/WUW behavior of memories
The HW dialect currently defines a `WUWAttr` to capture the
write-under-write behavior of memories. FIRRTL memories also specify a
read-under-write behavior, which suggests that we also need a `RUWAttr`.
Since these attributes talk specifically about the behavior of
sequential circuit elements, it seems sensible to move them into the Seq
dialect.

This commit adds a `RUWAttr` and a `WUWAttr` enum to the Seq dialect and
updates the FIRRTL dialect and `HWMemSimImpl` lowering to work with
these new attributes. The `WUWAttr` in HW is removed.

This change is a preparation step for later adding a `seq.firmem`
operation to capture the semantics of FIRRTL memories more precisely at
the HW dialect level in Seq.

Changes to tests are related to teh new `RUWAttr` being encoded as `i32`
instead of the manual cast to `ui32` we had before.