[FIRRTL][LowerXMR] force/release handling. (#4991)
Presently, we replace ref.resolve's with `sv.xmr.ref` operations that are cast to the referenced FIRRTL type.

Add support for force/release in a similar manner.

For force/release, don't replace but keep these around for lowering to HW in LowerToHW (#4992).  Instead, replace the "destination" reference operand similarly with a `sv.xmr.ref` cast to the reference (rwprobe) type.