[Handshake] Remove cmerge return type inference
This commit removes the InferTypeOpInterface from ControlMergeOp because
it no longer made sense w.r.t. the variable type of the index result.
Additionally, when parsing a control_merge with an index type different
than "index", it would raise an error because of the mismatch between
the parsed and inferred types. Example code that would trigger it below.

```
handshake.func @test(%arg1: i32, %arg2: i32) {
    %result, %index = control_merge %arg1, %arg2 : i32, i64
    return
}
```

The commit also adds a custom builder for ControlMergeOp to allow
creating the operation without explicitly providing return types. The
builder defaults to an Index type for the op's second result.