[FIRRTL] Add `IsTagOp` to test enumeration variant
This adds an `IsTagOp` which tests the active variant of an enumeration
for a specific tag and returns 1 if they match, 0 otherwise.  This op
will be used to lower match statements in to a mux-like format.  This op
should only be used internally in the compiler, and there are no plans
to expose this operation in the FIRRTL frontend language.