[FIRRTL] Add LowerMatches pass
The lower matches pass transform match statements into when statements.
This allows them to be handled by ExpandWhens, including all the fun
things it does such as initialization checking.  When lowering to when
statements, the final case handled by the match is transformed in to the
"default" branch and the enumeration tag for this case is not explicitly
checked anymore.

The new pass was inserted in to the firtool pipeline right before
ExpandWhens runs.

To pass flow checking, we needed to handle teach the `foldFlow` function
what to do when the block argument belonged to a match statement.
Although the flow should always be `source`, it is returning the
accumulated flow in case we relax the restriction that enumeration types
are always passive.