[Handshake] Remove handshake.select
I introduced handshake.select back in the day to distinct it from handshake.mux. What i somehow failed to realize is that `handshake.select` has always lowered to a unit-rate actor, which completely removes the need for the operation. The fix all along seems to just lower it to a comb/firrtl mux with unit rate semantics.

Furthermore, looks like there was a bug in the lowerings, wherein the operands of `arith.select` in standard to handshake was being swapped (hence the change in max.mlir, which was previously incorrect).

Handshake to FIRRTL is... eghh... could have done a "nicer" implementation which factors out the join logic, but i can't be bothered - this pass is supposed to be deprecated, anyways.