[LowerTypes] Preserve rwprobes, forceable.
Don't lower these as much as possible.

Keep the always-lower-(non-rw-)probe behavior
to workaround the memtaps / 4479 issue.

Also, if we need to lower the source operand
to a force statement, split it up as well.

Some sharp corners here re:references to non-passive things,
can be handled similarly to the force situation-- just
reconstruct or cat+cast the elements back.