[LowerTypes] Tweak how choosing whether lower ref.send.
Undo change that used the input argument, which was motivated by
ensuring a ref.send of a non-passive would be lowered since
we always lower non-passive types.

Instead, always look at the ref result for determining this,
so that ref chains are consistent.

We can rebuild the aggregate into the send as needed.

As-is we always expand read-only references anyway,
so this just ensures the send is also split apart
using same logic splitting up ref.define/etc of probe's.