[OM] Add initial implementation of Evaluator.
The Evaluator is responsible for evaluating the object model according
to the semantics defined by the rationale. It can be constructed from
an IR module, and Objects can be instantiated by supplying a Class
name and actual parameters.

Under the hood, it implements the small expression grammar outlined in
the rationale document. The entire grammar defined so far is
implemented in this initial version. Instantiations are currently
evaluated eagerly, which precludes cycles.

Objects are defined using a variant that is either a pointer to a
simple Object struct, or an typed Attribute for primitives. The fields
of an Object struct are stored in a simple map.

This is sufficient for an initial implementation, but lots of
opportunities exist to improve the evaluator, Object representation,
and storage.

Unit tests have been added that exercise all failure and success
modes.