Revert Alternative Connect/Invalidate Parsing
This reverts the following commits:

  - 5d371c3ce3d4faa6f3a8739f7024918e911c5e56
  - 0bb6ec6212f61330aba457e29e8cc91c2f009687

This has problems with situations where the leading token is either a
keyword or an identifier.  E.g., these two commits will cause errors on
ambiguous things like:

  connect is invalid

These will also cause errors on simpler cases like:

  connect <= foo
  invalidate <= bar

These need to do some checking to make sure that the token is not an
identifier before they proceed with assuming that it is the alternative
syntax.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>