[Seq] Add clock gate operation
Add a `seq.clock_gate` operation that represents a safe, glitchless
gating of a clock. It isn't totally clear to me which dialect this op
best fits into, but given that it interacts with clocks and contains an
internal storage element to avoid glitching, Seq seems like a good fit.

This commit merely adds the op, canonicalizations, and basic tests. It
isn't used or lowered yet. I imagine that the `LowerSeqToSV` pass would
be able to lower this clock gate either to an extmodule or to an inline
simulation model later.