[ExportVerilog] Remove uneeded parenthesis around UnionCreate
When no padding is reqiured for UnionCreate, we emit the expression
directly as the RHS of an assignment.  Since we always spill UnionCreate
to a wire, we don't need parenthesis around an expression on the RHS.
This PR switches the precedence of that case from `Selection` to
`LowestPrecedence` to elide it.

As an example, we used to emit:
```verilog
assign out = (in + in);
```
but we now emit:
```verilog
assign out = in + in;
```