[FIRRTL][ExpandWhens] Use port locs when emitting init errors (#5092)
Before this change, an error on a port used the location of the defining
module, instead of the location of the port, as this code was written
before block arguments had locations associated with them. We now emit
good port locations from Chisel, and reliably preserve them throughout
the pipeline, so it makes sense to use the port's location to emit the
error.

This also changes how we emit errors for regular operations. Since this
error message is expected to be user facing, it is better if we do not
dump the IR for the operation.