[FIRRTL] Parse literal identifiers
Add parsing support for FIRRTL Specification 3.0.0 _literal identifiers_.
This is a mechanism to encode an identifier that has a leading numeric
character (or any identifier if a user so chooses).  E.g., this enables
modules like the following to be parsed:

    module Foo:
      input `0`: {`1`: UInt<1>}
      node `2` = `0`.`1`

This is added for much easier lexing to avoid ambiguity on what is a
number vs. an identifier.  E.g., in the above "0.1" requires context
sensitive lexing whereas `0`.`1` is trivial.

In implementation, the tokens are unmodified, with the leading/trailing
backticks, to preserve existing indentation computations, but the names
are stored without the leading/trailing backticks.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>