[FIRRTL] Fix incorrect getSingleConnectUserOf when > 1 connect.
In particular, this caused a number of folder/etc users of this
method to erroneously believe there was a single driver for
the value if there is a connect AND a strict connect.

(The main one, canonicalizeSingleSetConnect, is safe because
it also conservatively checks the destination has a single use)

Also clarify this only returns the single connect if it's strict,
returning null if finds any non-strict connect driving the value.