[FIRRTL] Add Export of Force/Release Ops
Add force, force_initial, release, and release_initial to the FIRRTL
exporter.  This requires doing some "raising" of non-true predicates on
the initial ops.  This can be a source of "pseudo-canonicalization" where
the initial ops will converge towards a when representation which is
pretty high level.  It would be reasonable to think of a way to improve
this via a pre-pass in the exporter that does conversion to the `when`
representation and then tries to group common `when` blocks.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>