[Arc] Dedup: generalize to CallOpInterface (#5161)
Currently, the Dedup pass only considers StateOp as call operations to arcs because this was the only legal way to call an arc until recently. Because there is no additional check that no other operation calls an arc, the pass fails because it potentially dedups an arc that is called by a CallOp which is not updated accordingly. This generalizes to operations that implement the CallOpMutableInterface. It is assumed that all operations that can call an arc, implement this interface.