[arcilator] Convenience flag for printing debug info
To print debug info in MLIR output, the AsmPrinter CLI option mlir-print-debuginfo could be used. However, for printing debug info in the LLVM output, a pass to add the necessary DILocation attributes as to be called first. To make this process easy to use and uniform no matter whether MLIR or LLVM is printed, a print-debug-info CLI option is added