[FIRRTl] Add a type interface for querying phase compatibility
Phases separate the values in a firrtl program into two worlds, the property
world and the hardware world. Eventually, a firrtl program will be compiled
into two artifacts:

1. The hardware design (verilog)
2. The metadata (om dialect mlir bytecode)

The property phased values will be erased from the hardware design and only
exist in the metadata artifact, while the hardware phased values will be erased
from the metadata and only exist in the hardware design.