[WireDFT] Add support for wiring optional clock div bypass signal. (#5219)
If a signal is annotated as the clock divider bypass signal, wire it to clock gates that have the expected port.

Like test_en wiring the port number is hardcoded, but unlike test_en this ignores the "bypass" port if that port doesn't match the expected bypass port name/dir/type.

Warning is emitted if looks compatible but name doesn't match and bypass signal annotation is found.