[OM] Add Python API for ClassType of an Object.
This adds a type property to Objects, which returns the Object's
ClassType via the new CAPI. It also adds a ClassType Python class
definition, so we can create Python objects representing the
ClassType. No test is added, because Objects are not directly exposed
from the Evaluator Python API, but an upcoming change will make use
ofthis functionality.