[Pipeline] Add `pipeline.stage` op and rename others
This PR adds a new operation, `pipeline.stage` to the `Pipeline` dialect.
This operation is used to make stages explicit - that is, a region-defining
operation with inputs and outputs (much like the `loopschedule.pipeline.stage`).
The motivation for this is that once a `pipeline.pipeline` has had registers
materialized, a representational change to make the stages explicit allows
for much cleaner RTL lowering whenever stages are _not_ to be lowered to
hardware directly inline in the parent `hw.module`.

To (hopefully) avoid confusion, this commit also renames some existing
operations in the dialect:
* `pipeline.stage` becomes `pipeline.ss` - a *stage separating* operation.
  ... which is really what it is - it doesn't define a stage, it merely separates
  them.
* `pipeline.stage.regs` becomes `pipeline.ss.reg` - as above, it separates
  stages, but also defines the registers that are formed across the
  boundary.

A follow-up commit will rework HW lowering to be `pipeline.stage`-based.

doesn't really work

work