[IMDCE] Remove dead hierpath and annotations in IMDCE (#5226)
This PR adds functionality to find and remove dead hierpaths and weak referencing annotations in IMDCE. 
Even though it's not explicitly defined yet, some annotations (for example OMIRTracker)
are expected to be removed when annotation targets are dead. In many cases OMIRTracker 
is represented as non-local annotations associated with hierpaths which are users of inner symbols. 
So this PR imports some of InnerSymbolDCE  logic into IMDCE and aggressively removes inner symbols.