[Pipeline] Add optional `stall` signal to pipelines
This signal is intended to connect to all stages within the pipeline, and is used to stall the entirety of the pipeline. It is lowering defined how stages choose to use this signal, although in the common case, a `stall` signal would typically connect to the clock-enable input of the stage-separating registers.

Future PRs will implement hardware lowerings - blocked by #5234.