[OM] Add Python support for Objects in Fields. (#5276)
This adds support for Objects in the Fields of other Objects. This is
handled in the pure Python layer of the Evaluator API, which is
enhanced to support recursively instantiating Objects. The
functionality for creating a dataclass instance from an instantiated
Object is factored out into a new helper. In the case an Object field
is itself an Object, the new helper is called recursively with the
child dataclass and the child Object.

A simple integration test for this behavior is added.

This uncovered undefined behavior in the previous use of shared
pointers, so this patch depends on PR #5275.