[FIRRTL][ExpandWhens] Fix performance with large blocks
We need to make sure that we do not doubly process drivers when a wire
is driven in both the `then` and `else` block, so we were removing the
driver from the `else` block's driver map.  Since we are using
MapVectors to record drivers, which is needed for deterministic error
messages, deleting a single driver the `else` block's driver map scaled
linearly with the number of drivers.  Instead of removing it from the
`else` block's driver map, just check if it was also in the `then`
block's driver map.