[ETC] Enable extracting registers similar to instances.
The existing logic for extracting instances is made generic by moving
most of the helpers to work with Operation *, and templating the
method based on the concrete Operation type to extract. There are a
lot of changes in variable names and comments to reword things in
terms of operations instead of specifically instances, but the logic
is otherwise unchanged.

In order to enable the same generic logic to extract registers, the
pass has been moved earlier in the firtool pipeline, so it can work at
the seq.firreg level of abstraction, which behaves just like an
instance. The dataflowSlice method was updated to stop dataflow slices
at any of the Seq dialect ops.

Finally, a new call to the newly-renamed addOperationsToCloneSet was
added for the seq::FirRegOp case.

A test case is added that demonstrates registers that feed test code
and only test code can be extracted, including situations where test
code registers feed other test code registers. Registers that feed
both test code and design code are not extracted.

This logic is well-vetted in the instance case, so I think that should
be fairly low risk. My only other concern is if we will encounter
surprises by running the pass earlier in the firtool pipeline, but
looking at the intervening passes that had previously been run before
it, I don't have any concerns.