[IMCP] Mark overdefined for the operations with large number of operands (#5291)
We don't mark overdefined if there exists an operand with unknown state. This was required to get optimal results, but it also means that we have to try folding every time operand state was changed. It could cause regressions for operations with large number of operands. Actually in the internal designs there are operations with 65536 operands which caused regression in IMCP. This PR fixes the issue by just giving up such operations, especially when there are more than 128 operands since the success rate of folder should be much less. Maybe we can improve by specializing a folder for IMCP so that we don't have to get all operand lattice states.