[Pipeline] Refactor pipeline dialect to be block-based
This commit refactors the pipeline dialect to be block-based. This
brings a major representational change in the form of:
1. The pipeline is no longer defined by a lexical ordering of operations
and insertion of `pipeline.stagesep` operations to separate stages.
Instead, pipeline stages are defined by blocks.
2. Control flow between blocks are defined by `pipeline.stage` operations.
3. Like in the current version, the pipeline can exist in register
dematerialized and materialized forms. In the dematerialized form, stages
(`Block`s) have no arguments. In the materialized form, stages have
arguments.
4. It is the `pipeline.stage` operations which infers whether to register
a value or pass it directly (i.e. a wire) to the next stage.

The motivation for this change is to improve the hierarchy of the IR,
instead of relying on lexical ordering. This change also allows for
more natural traversal of stages (`Block`s), as well as dataflow analysis
of the pipeline, which now is analogous to control flow analysis.
The only slight drawback of this change is that it slightly complicates
adding new pipeline stages, seeing as one has to explicitly update
the control flow of the pipeline. This is a minor drawback, seeing as
this is also how things work in the software world, and is easily addressed
by helper methods.

Likewise, this change also removes the (old) `pipeline.stage` operations,
which mainly were introduced to facilitate lowering. This is no
longer needed with the block-based pipeline, seeing as stage in- and
outputs are clearly denoted by block inputs and `pipeline.stage` operations.